\name{rgRank}
\alias{rgRank}
\title{Bucketizing variables}
\description{
  Takes a continuous variable and produces a new score n-tile rank using the empirical cdf.
}
\usage{
rgRank(data, nTiles )
}
\arguments{
  \item{data}{is a vector of non-missing numerics.}

  \item{nTiles}{Gives the number of buckets.}

}

